+++ src/TinyGsmGSMLocation.tpp
@@ -174,15 +174,15 @@ class TinyGsmGSMLocation {
     }
 
     // init variables
-    float ilat      = 0;
-    float ilon      = 0;
-    float iaccuracy = 0;
-    int   iyear     = 0;
-    int   imonth    = 0;
-    int   iday      = 0;
-    int   ihour     = 0;
-    int   imin      = 0;
-    int   isec      = 0;
+    float   ilat      = 0;
+    float   ilon      = 0;
+    float   iaccuracy = 0;
+    int16_t iyear     = 0;
+    int16_t imonth    = 0;
+    int16_t iday      = 0;
+    int16_t ihour     = 0;
+    int16_t imin      = 0;
+    int16_t isec      = 0;
 
     ilat      = thisModem().streamGetFloatBefore(',');  // Latitude
     ilon      = thisModem().streamGetFloatBefore(',');  // Longitude
